Copy-on-write write locks in the cache
This commit introduces a copy-on-write mechanism into the cache. This
creates a new entry on cache_lock which can be written to, and frees the
entry on cache_unlock. This requires some minor changes to the cache
API.

Note that the cache_test is currently failing on pins, but everything
else is passing.