Remove cache_(un)share
This commit removes the cache_share and cache_unshare functions. These
were used in the memtable, and have been replaced with a simple page
allocation.

The cache_share and cache_unshare functions were designed to allow a
disk address to alias another disk address; i.e. after a share, the
second disk address would route to the same entry as the first. However,
due to the disk address checking done in cache_get, this alias didn't
work and a get to the alias would stall until cache_unshare was called.