(#106) Add tests for kvstore_iterator() with non-NULL start key
This commit adds couple of small test cases to exercise the
kvstore_basic_iter_init(), with a non-NULL start_key, to
validate that we are correctly positioning the start scan
on the requested key. Also, verify that if the specified start_key
does not exist that the scan is positioned on the 1st key,
if any, after the specified start-key. If start_key is
beyond max-key in data set, then iter_init() should
basically end up not finding any rows.

Minor miscellaneous cleanup:
 - kvstore_basic_iter_deinit() now takes address of iterator,
   so it can be Zero'ed out before return.
 - Mildly refactor existing test cases to consolidate rc handling,
   and inject PASS / FAILED message lines (for consistency).