Support managing output file handle(s) for platform_log()
This commit introduces a small infrastructure to manage the file
descriptor where output from platform_log() will go to. Currently,
this goes to stderr. This change provides the ability to swizzle,
in future, the output fh / fd for this interface to either /dev/null,
or to some user-specified logfile, depending on how the configuration
is setup.

This change is introduced initially to support CTests. The goal is
to leverage this fh/fd swizzling while running CTests to redirect
info-msg outputs off of stdout/stderr, so that these messages do
not clutter up stdout when running tests.