Support managing output file handle(s) for platform logging fns.
This commit introduces a small infrastructure to manage the file
descriptor where output from platform_log() and platform_error_log()
will go to.

Add platform_init_log_file_handles(), defined using the
"__attribute__ ((constructor))" feature to intialize the global file
handles to stdout / stderr, respectively.

This change provides the ability to swizzle the output fh for each
interface to either /dev/null, or to some user-specified logfile,
depending on how the configuration is setup.

This change is introduced initially to support CTests. The goal is
to leverage this fh swizzling while running CTests to redirect
info-msg outputs off of stdout/stderr, so that these messages do
not clutter up stdout when running tests.