(#218) Rework body of platform_assert() to avoid a function call.
This commit does minor rework to the body of the platform_assert()
macro to avoid a function call. In most cases, we expect the assertion
to succeed. This rework changes the body to a ternary so we check for
the common-case first, and only then invoke the function implementing
the extra functionality of the assertion checks.

Define platform_assert_false() with 'noreturn' attribute. Adjust tests
accordingly.

Rebase v/s main and fix fprintf() bug:
Tidy-up the body retaining the dead-code fprintf(), which is still very
useful to cross-check that print arguments match print formats in the
message. Fix nasty bug in definition of ":"-part of ternary, which
caused the supposed-to-be-dead-code fprintf() to always execute.

In misc_test, add test_platform_assert_msg() which reduces to just
verifying that the user-supplied message is bring printed correctly.