(#227) Refactor ctest_main() and related C-code out of ctest.h to main.c
This commit relocates and rearranges a whole bunch of CTest code that
was defined in ctest.h to main.c . This way, the flow from main() ->
ctest_main() and minions is better seen through one source file.
Minor rearrangements of functions done in this file to keep things
in a somewhat logical order.