(#274) Support --list arg to list unit-test suite-names & test-case names
This commit adds the --list argument which can be used to list all the
suite-names that can be run from the unit_test binary. It can also be
used to list the test case names for a specific suite-name, or for a
standalone unit-test binary. Overhaul the unit-test README.md docs to
reflect the current state of unit testing build and test execution
interfaces.