Fix bug that causes 'setup' to be printed; Other tightening.
There was a flaw in the processing logic of
print_test_suite_names() which was skipping placeholder
begin / end test suite definition entries that the CTest
framework builds. These entries could appear anywhere in
the list, so skipping has to be done at the start of the
for()-loop. [ This is how it's being done in other such
canonical loops around test suite definition arrays. ]

- Added a small optimization for the case when
  'unit_test --list suite-name' is run. Identify the
  start of the requested suite-name. Once we move past
  that to a diff suite-name, bail out early. For longer
  lists of suites, this will make a small difference in
  performance of --list processing.

- Other minor coding convention cleanup.