(#282) Revert from Ubuntu 22.04 to v20.04, to resolve CI-incompatibilities
Recent upgrade to Ubuntu 22.04 _seems_ to run into some issues
due to impedance with CI-test infrastructure. This is suspected to be
the cause of some tests not working on CI, or failing to execute due
to permission issues. This commit reverts the changes done in SHA e7d0ae0
(Use Ubuntu 22.04 LTS in CI). We go back to Ubuntu 20.04 "focal"
version, using default gcc (v9.3.0) and install clang-13, plus
clang-format-13, for CI runs.