fix dependency issues in Makefile (#284)
Fix several issues with dependencies in the Makefile.

It turned out there were several pitfalls that needed to be avoided. Read on for what I learned.

I had been observing that sometimes a make would succeed even though it had not recompiled all the files changed since the last build. The first layer of issue was rules of the form:

some_target: dependency1.o
  ld -o some_target dependency1.o dependency2.o
Note that some_target really depends on two .o files, but make is told about only one of them. Doh!

The fix for this is to write all recipes to refer to the dependencies in the rule and to never list any inputs explicitly in the build recipe. That way an omitted dependency will fail as an error when you attempt to run the recipe. The Makefile now does this. The only recipes are on lines 157-169.

The second pitfall, which I suspect is why some violations of the first rule were introduced, was the use of convenience targets in rules. For example, a convenience target is fine by itself:

bin/foo: ...
  ....

# a shorthand way to build bin/foo
foo: bin/foo 
But convenience targets should not be used as prerequisites of other targets:

# OK: bar depends on its real prereq
bin/bar: bin/foo

# WRONG: baz depends on the convenience target
bin/baz: foo
There are two problems with this. First, baz will always get rebuilt unnecessarily, since foo will never exist. Second, if there was a generic pattern recipe for, say, building a %/baz from a %/foo, then make will not find it. As a result, this target may be written with its own one-off recipe. And that recipe may be wrong or become wrong over time, since it is less likely to be maintained with the generic recipes.

Another problem I encountered is that the way ctest automatically discovers all the tests in the binary is prone to silently omit tests when you omit a .o in the linking phase. Briefly, ctest creates a special section of the binary that contains an array of data structures, each of which describes a test. So, if you add another .o file on the linking command line, then its tests get added to the special section, and its tests get run automatically.

As a result, omitting a .o on the linking command for a ctest executable results in a successful executable that simply doesn't run the tests. In other words, it happily reports that everything is ok even though some or all of the tests are missing. So we need to be very careful about ctest executable dependencies.

Finally, to support building individual ctest executables for each subsystem, we need to list the dependencies of each mini unit test by itself. I tried using pattern rules to specify common elements of their dependencies, and then concrete rules for each test's specific dependencies. This does not work, because make will discard any dependencies from pattern rules as soon as it sees a concrete rule for your target. The work-around is to generate concrete rules algorithmically. See lines 198-201.

I haven't fully fixed the logic for switching between release and debug builds. I am concerned that the current approach is fundamentally incompatible with make. I would propose to fix it by moving bin, lib, and obj under a build directory, and having two different build dirs, e.g. debug-build and release-build. But I wanted to get feedback before doing that.