Bugfix for Type Assertion in `trunk_async_lookup`
Fixes a bug where an assertion that `type` is not update would be hit in
`trunk_async_lookup` on function exit. This could happen when an update
message is found, but there is still further IO required to finish the
lookup.

This fix moves the relevent code, including the assertion, into the
`async_state_end` state.