(#371) Pre-stage CI-jobs' script changes to work with Makefile changes.
Upcoming PR #332 changes Makefile targets to build the library. Those
changes will require changes to the commands invoked by CI-jobs to build
and test the library. This commit pre-stages those CI-build script
changes, so that the upcoming Makefile changes can be tested independently
before they are merged to /main. (This change depends on PR #372, which
adds minimal support for 'help' target to Makefile.)