(#252) Set of example C-programs, to demonstrate use of SplinterDB APIs.
This commit adds a small number of C-sample programs, to demonstrate use
of SplinterDB core APIs. Associated changes to build this set of stand-alone
programs applied to Makefile. A small new "test-case" is added to
test.sh to always execute these programs as part of default testing set.
Associated docs files are also updated to direct the user to these programs.

Add a test-case to execute generated example programs.

- Update docs to href-link to example C programs.
- Minor edits to indent Makefile code.