(#305) Add instrumentation to track assertion in clockcache_get_internal()
splinter_test --perf, and some other tests, are randomly tripping up a
debug assertion in clockcache_get_internal(), which reports that we
are trying to cache a page from an unallocated extent. This problem is
not reproducible. This commit improves the assertion message to report
addresses involved, and also calls-out to a print method to dump out
allocated extents bitmap. (This should only occur very rarely, if at all.)