(#305) Add instrumentation to track assertion in clockcache_get_internal() (#401)
splinter_test --perf, and some other tests, are randomly tripping up a
debug assertion in clockcache_get_internal(), which reports that we
are trying to cache a page from an unallocated extent. This problem is
not easily reproducible outside CI-runs where it is seen to occur sometimes. 

This commit improves the assertion message to report addresses
involved, and also calls-out to a print method to dump out allocated
extents bitmap. (This heavy-dump should only occur very rarely, if at all.)