(#120) Handle invalid user MSGs in 'btree_insert' and 'btree_pack'
In `btree_insert` and `btree_pack`, invalid user-type messages are not
handled properly. This change makes those functions return an error
when they receive invalid user-type messages. An error that
`btree_insert` returns is forwarded to an application. On the other
hand, when `btree_pack` returns an error, the program will be
asserted.