(#424) Add instrumentation and extra validations to narrow down trunk assertion.
In YCSB runs against FDB + SplinterDB, we are running into an assertion
complaining that some key-value stats are off after a compaction event.
This commit adds extra diagnostcs / tracing to troubleshoot the code-flow
of this error. Add print methods to dump out all fields of related
structures.

Also bracket call to trunk_replace_bundle_branches() with a call to
trunk_verify_node() before after this fn-call. This is to see if the
error can be found before replace_bundle is called.

With these changes, we are now getting extra [useful] diagnostics for this
YCSB failure, but more investigation is needed.