Reduce the usage of malloc
Previously, there are two locations using malloc frequently. One is a
set of `tictoc_rw_entry`, and the other is `interval_tree_node`. These
make a lot of number of malloc, resulting in a performance degradation.
Now, a transaction creates one `tictoc_rw_entry`, and this instance will
be used for a lock tree node.

FIXME: The current code need to be clean. For example, `struct
tictoc_rw_entry` is in `lock_table.h`.