(#444) Refactor chunks of test.sh into functions for code-reuse (#445)
The script test.sh has several inline calls to run different
tests. Many of these are interesting cases to be run with
an upcoming feature to run Splinter using allocated shared
memory segments.

This commit rearranges and refactors inline calls to run
diff tests into new functions that will now execute a
collection of related tests.

Examples of new functions added are:
 - run_fast_unit_tests(), run_slower_unit_tests()
 - run_splinter_functionality_test(), run_splinter_perf_tests()

and so on. No changes are made to the set of tests that will
be executed. This refactoring lays down the stage to re-execute
relevant set(s) of tests using shared-memory.