Rework io_context_t handling, make it an array of thread-specific IO context.
This commit sets up the stage to support fork()'ed or other OS-processes
running with --use-shmem option, where each process will [in future]
masquerade as a Splinter thread. A core change needed to move to that
execution model is to support thread-specific IO-context structures.
Otherwise, if another OS-process tries to do IO using AIO-context
established by the main thread (i.e. by the eprocess that started up
SplinterDB), we will immediately run into hard IO-system call errors.

This commit:
 - Performs an io_setup() for each thread / process
 - Manages this AIO-context tightly bound to Splinter's thread context
 - Does required book-keeping to keep this IO-context state kosher
   in conjunction with thread registration / de-registration.
 - Updates existing io_apis_test to deal with thread-specific IO
   context handles.

An alternative could be localize this change-in-behaviour (of setting
up thread-specific IO-context structs) only when the process-model of
execution comes around. That execution model requires configuring
SplinterDB with shared-memory support. But, just by looking at
--use-shmem (or corresponding config setting), we cannot be sure that
the process-model will be used or if we are just re-running rest of
the test suites with shared-segment enabled. So, without trying to
further complicate this choice-making, with this commit we will always
set up thread-specific AIO-context structures.