Fix platform_get_heap_id() to return stashed away heap-ID handle.
Accessor method platform_get_heap_id() is (was) incorrectly returning
NULL even when Splinter was configured to use shared memory. This
presents a problem when, in some code path, in the running code, this
API is called to allocate some memory. Result is that even while most
structures are allocated from shared segment, some structures will
be allocated from malloc()'e code.

One place where this occurs is in platform_buffer_create(), which
allocates the 'buffer_handle *bh' using allocated memory.
In a threaded usage, this is not a problem. But when multiple processes
use Splinter with shared memory, we get instablity which manifest
in unlikely ways; such as finding a valid BTree page as unreferenced
in cache's ref-count array.

This commit 'fixes' this issue by tracking 2 new globals declared
in platform.c . Whenever shared-memory creation sets up the platform
heap handle / ID, those are saved of in these globals and the heap-ID
is returned via platform_get_heap_id().