Fix bug in mmap() flags that is causing basic instability.
Basic unit-tests exercised through Pg psql clients is showing some
instability. Steps are:

- Insert couple of rows from psql client.
   - Examine that these inserts are seen on BTree pages.
- Shutdown Postgres, which invokes splinterdb_close().

While unmounting SPlinter, we run into an assertion reporting that
the BTree iterator found 0 rows.

Assertion failed at src/trunk.c:3207:trunk_memtable_compact_and_build_filter(): "(req.num_tuples > 0)". num_tuples=0, generation=0, tid=0

Stack is:

 #5  0x0000559e8356c840 in platform_assert_false (filename=0x559e837b51fb "src/trunk.c", linenumber=3207,
        functionname=0x559e837b9600 <__FUNCTION__.92> "trunk_memtable_compact_and_build_filter",
        expr=0x559e837b5c21 "(req.num_tuples > 0)", message=0x559e837b5bf8 "num_tuples=%lu, generation=%lu, tid=%lu\n")
        at src/platform_linux/platform.c:353
    #6  0x0000559e835782b5 in trunk_memtable_compact_and_build_filter (spl=0x7f3ccebe7640, generation=0, tid=0)
        at src/trunk.c:3207
    #7  0x0000559e835794d0 in trunk_memtable_flush_internal (spl=0x7f3ccebe7640, generation=0) at src/trunk.c:3487
    #8  0x0000559e83579555 in trunk_memtable_flush_internal_virtual (arg=0x7f3ccec388b0, scratch=0x7f3cce5095e8)
        at src/trunk.c:3505
    #9  0x0000559e83596926 in task_group_perform_one (group=0x7f3cce507040) at src/task.c:696
    #10 0x0000559e83596aad in task_perform_one (ts=0x7f3cce505b40) at src/task.c:722
    #11 0x0000559e83596484 in task_perform_all (ts=0x7f3cce505b40) at src/task.c:624
    #12 0x0000559e835873dd in trunk_prepare_for_shutdown (spl=0x7f3ccebe7640) at src/trunk.c:7325
    #13 0x0000559e83587923 in trunk_unmount (spl_in=0x7f3cce4da9f8) at src/trunk.c:7420
    #14 0x0000559e8356a821 in splinterdb_close (kvs_in=0x559e838ce480 <Splinter>) at src/splinterdb.c:623

===
The issue is that some parts of structures, e.g. buffer cache, are allocated using
mmap(). The flags for this were MAP_PRIVATE, which means this memory is only
accessible to the main process that set up Splinter. All child threads work on
a COW-version of this mapped memory. So the changes done by the child process to
the BTree in the buffer cache are not visible to the parent process.

This commit fixes the issue, adds diagnostics in various places to better report
state of failing assertions and a test case that demonstrates the problem w/o the
fix in mmap() call.