Fix bug in mmap() flags that is causing basic instability.
Basic unit-tests exercised through Pg psql clients is showing some instability.

The issue is that some parts of structures, e.g. buffer cache, are allocated using
mmap(). The flags for this were MAP_PRIVATE, which means this memory is only
accessible to the main process that set up Splinter. All child threads work on
a COW-version of this mapped memory. So the changes done by the child process to
the BTree in the buffer cache are not visible to the parent process.

This commit fixes the issue, adds diagnostics in various places to better report
state of failing assertions and a test case that demonstrates the problem w/o the
fix in mmap() call.