Allocate fingerprint array from shared memory; not from malloc()
In previous versions, while developing support for shared segment,
certain allocations were changed to revert to using malloc()-based
allocation. The NULL_HEAP_ID indicator was provided for this.

btree_pack_req->fingerprint_arr was one such consumer changed to
use NULL_HEAP_ID. This seems to work fine if we run with pthreads
but when multiple child processes do heavy inserts, we run into
assertions / signals. This seems to be because one process may have
done the allocation, and registers a task to be completed while
doing trunk build-filter. Another process may pick this up and
may end up eventually freeing this memory. Depending on how things
play out, we end up seg-faulting. Debugging individual processes
showed that we run into a hard-error from free() complaining that
the address is invalid. This is likely if the allocation was done
by some other process.

Reverting the use of NULL_HEAP_ID in:
  - btree_pack_req_init() and btree_pack_req_deinit()
  - merge_iterator_create() and merge_iterator_destroy()

seems to have addressed this issue. We
no longer get segvs or other fatal errors while the concurrent
processes are inserting. (We do get other errors from the main
process, but has to be debugged separately.)

One consequence of this change is that for very large inserts,
say 20M per process, we run into shmem-OOMs as there is no support
for real free from shmem.