Implement recycling, under spinlock, of large free fragments.
This commit reverts allocating of fingerprint array back from shared
memory. This is a big array, so we run out of shared memory very
easily. This commit adds special-case handling for large
fragments by tracking a small array of freed fragments. Upon a
request for a large fragment (>= 1MiB), we may recycle previously
freed large fragment to the requestor. The array of free memory
fragments is managed under a spinlock.

Fixed an imp bug in test code where we were not wait()'ing for
all child pids to complete.

With these changes, the test test_multiple_forked_process_doing_IOs
seems to be passing somewhat more reliably. We still are running
into some assertions occasionally that will need to be debugged.