Changes to run large insert stress tests with 63 threads.
We are seeing assertions in PG-integrated Splinter code when we
have more than, say, 50 or so concurrent psql clients firing
inserts. Assertion in laio.c to do with IO-context indexed
by 'tid' is failing:

OS-pid=1897373, Thread-ID=18446744073709551615, Assertion failed at src/platform_linux/laio.c:109:io_context_setup(): "(io->ctx[tid] == NULL)". ThreadID=18446744073709551615, ctx[tid]=0x7ffab9d7088    8

The assertion is valid, but the 'tid' reported seems to be a wrong
number. We expect this to be in the range of [0-63]. Somehow this
reeks of an uninitialized threadid variable.

This commit tries to repro this assertion, using existing unit tests.
But the attempts were unsuccessful. Some changes were needed in
configuration to run with --num-threads 63.

- Fix some off-by-1 checks in clockcache.c, around use of MAX_THREADS.
- Add field shm_large_frag_hip in shm-control block to track the
  highest-start address of large fragments that are being tracked.
  This is a light-weight optimization to reduce the # of searches
  thru the array where any generic memory fragment is freed.

With these changes, following ran successfully, and could not repro
above assertion:

$ large_inserts_bugs_stress_test --use-shmem --num-inserts 1000000 --num-threads 63 test_seq_key_seq_values_inserts_threaded