Minor adjustments for background thread support to work with Postgres integration.
Relocate num_bg_threads[] in splinterdb_config{} structure from the tail
of the struct to somewhere in the middle. For some strange reasons, the
start offset of this array is coming out differently when seen in
Pg-splinteram.c, and in splinterdb.c . Tried to debug this, but could not
figure out the real reason. Possibly due to compiler padding / alignment
issues in the two builds.

Deleted junk trace messages reporting on background threads' setup etc.

With these fixes, PG-integration boots again (default bg-threads are OFF).