Redefine 'bool' to 'unsigned char' on Splinter-side.
For Postgres-integration, Splinter's definition of bool as int32
throws-off some structure layouts. Specifically, a change being tried
is to add array of num_bg_threads[] to splinterdb_config{} struct. That
struct has a few 'bool' fields defined, which is defined as 'unsigned char'
in Postgres code base. This throws off the offsets of all fields in
this structure following the 'bool' fields.

We may need a longer-term resolution, but for now redefine bool as
'unsigned char' to get this integration moving forward.

Fix merge errors.