(#462) Prepend OS-pid/tid and platform thread-ID to assert message. (#471)
To better debug assertion failures, this commit extends the
assert message to prepend the OS-pid, OS-tid and thread-ID (index)
of the running Splinter thread that encountered the assertion.
Couple of existing test cases are mildly amended to cope with
these new fields in the assert message printed. Existing
debug_assert() also [naturally] gains this functionality as it's
a wrapper around platform_assert().