Trace list of large fragments tracked when SplinterDB shuts down.
We are running into shmem-OOM while allocating large fragments. It's
puzzling why the scheme seems to work, in general, but we are still
over-allocating shared memory for large fragments.

This commit adds diagnostic code to trace the set of large fragments
[still] tracked in the shmem control block. Provide a new global,
Trace_large_frags, and helper methods to turn this ON.

With this infrastructure, when we run large-inserts workloads through
Pg-integration, we do see some large fragments still "in-use" when
Splinter is shut down. This indicates some form of memory leak.