Fix memory leak in trunk_compact_bundle(). Free fingerprint array.
In a degenerate case deep inside trunk_compact_bundle(), when
"num_replacements == 0", we are freeing the trunk_compact_bundle_req
req structure, but are failing to free memory allocated for the
fingerprint array. Over large stress tests, this results in OOM
in shared segment, and diagnostics code shows that a large collection
of fingerprint array memory chunks are still "in-use".

The fix is to free the fingerprint array also. With this fix, we
are now able to run large concurrent insert workloads in Postgres
integration, w/o running out of shared memory. We no longer need to
crank-up shared memory as high as 16GB. Haven't retested this but we
should be able to run fairly high stress concurrent insert workloads
even with reasonable [2-4 GB] shared memory segments.