(#481) Fix missing free for fingerprint array in trunk_compact_bundle() (#483)
This commit fixes a missing free for memory allocated to the fingerprint
array. The problem was seen to occur in highly concurrent inserts stress
workload developed as part of concurrent POC-development and is difficult
to reproduce in standalone Splinter.

In trunk_compact_bundle(), for a specific code-flow when we found no newly
split siblings to replace with a new branch, we release the memory allocated
for the trunk_compact_bundle_req{} struct, and exit processing. But the
memory allocated (elsewhere, previously) for the fingerprint array hanging
off this req struct remains unfreed.

This fix was identified through trace diagnostics and code walk through.
And has been independently verified to fix this memory leak in the stress
workloads that were used to generate this error.