Delete newly added large_inserts_bugs_stress_test
Pre-review discussion indicates that we should find other ways to
enhance existing tests (e.g. functional/splinter_test) to provide the
coverage that the sub-cases in this test are written for.

Also, some sub-cases in this test are failing small # of background
threads are configured. This causes CI to fail. Withdraw this test so
the main enablement feature can move ahead while we rework test coverage
in the background.