(#511) Add fns to print 'size' as human-readable string w/ unit-specifiers
This commit adds couple of utility functions to snprintf(), in an output
buffer, the 'size' unit to a human-readable string with unit-specifiers.

- size_to_str() - Convert 'size' to a string in an output buffer
- size_to_fmtstr() - Same as above, using user-specified format-string.
   Useful to generate output enclosed in, e.g., '(%s)'.
- Add size_str(), size_fmstr() caller-macros to simplify calling these
  formatting functions. These macros declare on-stack buffers used to
  format the output string. size_str() provided by Rob Johnson, to greatly
  simplify the usage.

Add utility bytes-to-Units conversion macros. Add unit tests to exercise
these interfaces. Apply these utility fns in couple of stats-printing
and BTree print-methods, to display size values as human-friendly unit
specifiers.