(#511) Add fns to print 'size' as human-readable string w/ unit-specifiers
This commit adds couple of utility functions to convert 'size' unit to
a human-readable string with unit-specifiers, in an output buffer.

- size_to_str() - Convert 'size' to a string in an output buffer
- size_to_fmtstr() - Same as above, using user-specified format-string.
   Useful to generate output enclosed in '(%s)'.

Both conversion functions return the start of the supplied output buffer
after snprintf()'ing the size-value into it with unit-specifiers.

Add utility bytes to Units conversion macros. Add test cases to exercise
these interfaces. Apply these utility fns in couple of stats-printing
and BTree print-methods, to display values in units of bytes as a string.