(#342) Fix checks in trunk_config_init() to report correct arg-values.
This commit does a minor cleanup in trunk_config_init() to correctly
report the value for 'Maximum key size' reported when the checks
fail for an invalid configuration. We were overflowing the value
for an uint64 variable, leading to bogus values reported in the error.
Changed this function and other helper functions in test-code that parse
command-line arguments to return platform_status. Adjust the handling
of this failed return-value in diff test code. Add new unit-test to
verify that invalid config params correctly fail to init the trunk-config.