(#290) Replace use of 'find', 'grep' with Makefile built-in 'wildcard' etc.
This commit replaces all uses of shell 'find', 'grep' commands
in Makefile with built-ins such as 'wildcard', 'filter-out' etc.
This removes any potential dependency on shell commands for 'make'
invocations. A small change is to list the unit-test sources in
lex-sorted order when linking them into the unit_test binary.