Fix 'realloc' capability. Interface of writable_buffer_resize() changed. Stabilize tests.
This commit fixes a core bug in shmem-based functionality of 'realloc'.
We are supposed to copy over the contents of the 'oldptr' to the 'newptr'
allocated. This was broken in diff ways and is fixed by this commit.

This fix now enables running this test, in particular, which discovered
the bug:

$ /driver_test splinter_test --functionality 1000000 100 --use-shmem --key-size 8 --seed 42

General clean-up around writable_buffer_append() -> writable_buffer_resize()
to deal with 'realloc' requirements. This single fix basically stabilizes
the test execution. Online functional and perf tests in test.sh .
We are able to cleanly run the entire CI-suite of tests that are enabled
to run with --use-shmem option.