Add platform_mem_frag{}, passed to platform_free() to free memory objects
This commit starts to bring back the interface to platform_free() and
keep platform_free_mem() as an internal interface. Callers that
allocate memory fragment for an opaque type, say an array of n-bytes,
will now have to package their memory fragment's info in a new
platform_mem_frag{} struct. platform_free() is enhanced to recognize
an object of this type, and will do the necessary unpacking to
call the lower free methods.

Fix a bug in trunk_unmount() to free memory for stats-fragments.