shmem.c: Rename variables relating to large-fragment handling.
Upcoming PR #569 is bringing-in support for handling small
fragments. This commit renames existing variables, field names
and a few function names that deal with large-fragment support
to consistently use 'large' in the name. This clears the way
in the namespace for code changes coming from small-fragment
changes.

Some examples:
- struct shm_frag_info -> struct shm_large_frag_info
- E.g., shm_frag_addr -> frag_addr, shm_frag_size -> frag_size ...
- shm_frag_info shm_mem_frags[] -> shm_large_frag_info shm_large_frags[]
- shm_num_frags_tracked -> shm_nlarge_frags_tracked
- platform_shm_find_free() -> platform_shm_find_large()

NOTE: No other code-/logic-changes are done with this commit.