Robj/leaf split fix (#624)
Fix bug where a leaf split that resulted in a single leaf would cause an assertion failure in a subsequent compaction.

The cause of the bug was that the compaction code detected node splits by checking whether the upper-bound key of the node had changed since the compaction request was created.  However, in a leaf split into 1 leaf, the upper bound key doesn't change, but the leaf does get rebundled, meaning that the bundle in the compaction request is no longer live.  This meant that the compaction thought it was seeing a dead bundle in a node that had not split.  This should only occur if the bundle gets flushed from a parent to a child between the enqueuing of the compaction and its execution.  But that can only occur if the node is not a leaf.  Hence the compaction code asserted that the node was not a leaf in this case.

To fix the bug, we improve the way node splits are detected. Every node has a unique id in its header.  IDs change whenever a node is split, which makes detecting splits trivial.