Rewrite trunk for better cow and modularity
This is a complete rewrite of the trunk to make it copy on write, to enable it to garbage collect old nodes, to make the code more modular, to separate the on-disk representation of nodes from the in-memory representation, to enable nodes to be larger than a page, and to fix several liveness and concurrency bugs in the old trunk.

The code introduces several new components, including a new system for writing async code, a new vector data type, and new code for easily printing well-formatted tables.

The code also contains a slew of other bug fixes encountered while solidifying this code.